fix: misc issues related to refactoring
